RTの時にMeshとObjectがリンクできるようにデータを用意した

+++ core/framework/include/Subsystem/Render/Descriptor/DescriptorSetBase.h
@@ -28,7 +28,7 @@ namespace MyosotisFW::System::Render
 
 	protected:
 		void createDescriptorSet();
-		void buildSSBODescriptor(const uint32_t index, const uint32_t dataSize);
+		void buildSSBODescriptor(const uint32_t index, const size_t dataSize);
 		void buildASDescriptor(const uint32_t index, const std::vector<VkAccelerationStructureKHR>& handle);
 
 		RenderDevice_ptr m_device;
